Add ENABLE_AVAHI env var to control dbus/avahi usage in container
ENABLE_AVAHI decides whether dbus and avahi daemons run inside the
container or if the host's instances are reused. When set to 0,
/var/run/dbus and /var/run/avahi-daemon/socket must be bind-mounted
by the user from the host into the container.

+++ docker/run.sh
@@ -3,15 +3,17 @@
 # exist if any command returns a non-zero result
 set -e
 
-rm -rf /run/dbus/dbus.pid
-rm -rf /run/avahi-daemon/pid
+if [ -z ${ENABLE_AVAHI+x} ] || [ $ENABLE_AVAHI -eq 0 ]; then
+  rm -rf /run/dbus/dbus.pid
+  rm -rf /run/avahi-daemon/pid
 
-dbus-uuidgen --ensure
-dbus-daemon --system
+  dbus-uuidgen --ensure
+  dbus-daemon --system
+fi
 
 (/usr/local/bin/nqptp > /dev/null 2>&1) &
 
-avahi-daemon --daemonize --no-chroot
+[ -z ${ENABLE_AVAHI+x} ] || [ $ENABLE_AVAHI -eq 0 ] || avahi-daemon --daemonize --no-chroot
 
 while [ ! -f /var/run/avahi-daemon/pid ]; do
   echo "Warning: avahi is not running, sleeping for 1 second before trying to start shairport-sync"
